pyproject.toml
[tool.poetry]
name = "pixel-detector"
version = "0.2.0"
description = "Healthcare website pixel tracking detection system"
authors = ["Ming Yang"]
readme = "README.md"
packages = [{include = "pixel_detector", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.41.0"
playwright-stealth = "^1.0.6"
pydantic = "^2.5.0"
httpx = "^0.26.0"
rich = "^13.7.0"
typer = "^0.9.0"
dnspython = "^2.6.0"
boto3 = "^1.34.0"
tldextract = "^5.1.3"
setuptools = "^69.0.0"
awslambdaric = {version = "^2.0.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^24.3.0"
ruff = "^0.3.0"
mypy = "^1.8.0"
ipython = "^8.20.0"

[tool.poetry.scripts]
pixel-detector = "pixel_detector.cli:app"

[tool.poetry.extras]
lambda = ["awslambdaric"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "S", "B", "A", "C4", "PT", "Q"]
ignore = ["S101", "B008", "UP007"]  # S101: assert in tests, B008: typer defaults, UP007: Optional[T] for typer compat

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
